requirements.txt
pytest
paho-mqtt==1.5.0
pynput==1.6.8
mock
Flask==1.1.1
Flask-Migrate==2.5.2